Move stocktake delete navigation out of hook
